docs: centralize site base-path handling
